ControlBoardHwVersion2 20200313(Ver_2003.0)    修改点：A016设置为1时，A010变为256，257，258，应该是0，1，2

+++ ControlBoardHwVersion2_code/ControlBoardHwVersion2/Middleware/paramdatabase.h
@@ -145,7 +145,7 @@ extern uint8_t gu8_en_oprdelay;
 extern uint8_t gu8_modefix;
 extern uint8_t gu8_en_switchpan;
 extern uint8_t gu8_opr_rsrttmr;
-extern uint8_t gu32_intlck_deltmr; //version 0.7D of C file
+extern uint32_t gu32_intlck_deltmr; //version 0.7D of C file
 extern uint8_t gu8_close_oprset;
 extern uint8_t gu8_opr_cntres;
 extern uint8_t gu8_init_valueset;
